requirements.txt
yfinance
pandas
numpy
matplotlib
scikit-learn
requests
beautifulsoup4
lxml
plotdigitizer
PyPDF2
pytesseract
Pillow
ta
tqdm
python-dotenv
groq
openai
anthropic
streamlit
plotly
httpx
apify-client